Resolve external dependencies relative to package base.
For example, if we have `loadstar.require(‘lodash’);`

Then, in package.json, we can have:

```
“loadstar”: {
  “dependencies”: { 
    “lodash”: “./dist/lodash.min.js”
  }
}
```

Which will resolve to node_modules/lodash/dist/lodash.min.js

An absolute path will be rooted at our project root:

```
“loadstar”: {
  “dependencies”: { 
    “lodash”: “/vendor/lodash-mine.js”
  }
}
```

If you put false as the value inside the dependencies
object, then no warning will be output. Use this
if you are building a library that has external dependencies,
and allow the user of your library to add the dependency.